[3] The hero animation is not deterministic
When the hero moves - in the saved file named save - between the snake and the skeleton - s and e tasks -, the walking animation is not the same when going from s to e and when going from e to s. Sometimes, it also changes mid walk without changing the target.