[current-starting-time] Displaying the current starting time in taskoid screens WIP
"Hiding" and "Unhiding" the taskoid in the Game World still has to be implemented. The entity has to be signalled, when it was "hidden", so that it's "unhidden", when the next iteration's time has come. Actually hiding the taskoid can be separate into a separate task, since the correct behavior is not defined ATM - e.g. difference between tasks and projects.