[current-starting-date] Implemented hiding and unhiding reviving entities
A bunch of stuff had to be changed, so that I can test the reviving entities, so now monsters are hidden - until we get a killed monster sprite - and portals are kept on the "finished" animation. Honestly the work done on the monster hiding - well at least part of it - was not only unnecessary, but also will have to be changed once we get the killed monster sprite and want to make it work. But oh well... it is what it is.